Fix test timeout in CI
